First set of w4-wall warning fixes.

+++ library/include/makina/renderer/backend/bgfx/buffer.hpp
@@ -26,7 +26,7 @@ public:
     };
 
     type               type              ;
-    std::size_t        size              ;
+    std::uint32_t      size              ;
     ::bgfx::VertexDecl vertex_declaration;
   };
